requirements.txt
twilio
flask
pandas
openpyxl
google-generativeai
python-dotenv
pyngrok
sqlalchemy
google-genai
trycloudflare